+++ peer.h
@@ -22,15 +22,15 @@ protected:
     virtual void closeHE(uint32_t events) = 0;
 public:
     virtual ~Peer();
-    ssize_t windowsize = 65535; //(for http2) 对端提供的窗口大小，发送时减小，收到对段update时增加
-    ssize_t windowleft = 65535; //(for http2) 发送给对端的窗口大小，接受时减小，给对端发送update时增加
+    int32_t windowsize = 65535; //(for http2) 对端提供的窗口大小，发送时减小，收到对段update时增加
+    int32_t windowleft = 65535; //(for http2) 发送给对端的窗口大小，接受时减小，给对端发送update时增加
     virtual void clean(Peer *who, uint32_t errcode);
     virtual ssize_t Write(Peer* who, const void *buff, size_t size);
     virtual int showerrinfo(int ret, const char *) = 0;
     virtual int showstatus(Peer *who, char *buff) {return 0;}
     
     virtual void writedcb();
-    virtual size_t bufleft(Peer*);
+    virtual int32_t bufleft(Peer*);
     virtual void wait(Peer *who);
 };
